#79 - Refactor accessibility support out to helper class.
